Create Mini games
A series of games coded in C++ that emphasizes the usage of the fundamentals in programming. RayLib was also used in order to facilitate many of the physics and collisions used in the game. You must download the specific images and ray lib files to play the game as well as a C++ IDE.