Add files via upload
The game of pure strategy created for a coding competition. Members input their strategies into our program and go to head to head against another member. The game goes 100 times and the strategy that wins most often is declared the winner. The game of pure strategy is a card game. Search online for clarity.